separate AST
